move aggregate transformation code to dedicated package
