+++ src/main/java/at/ac/tuwien/kr/alpha/grounder/transformation/aggregates/AggregateLiteralSplitting.java
@@ -1,4 +1,4 @@
-package at.ac.tuwien.kr.alpha.grounder.transformation;
+package at.ac.tuwien.kr.alpha.grounder.transformation.aggregates;
 
 import org.apache.commons.lang3.tuple.ImmutablePair;
