+++ src/test/java/at/ac/tuwien/kr/alpha/grounder/transformation/AggregateOperatorNormalizationTest.java
@@ -10,6 +10,7 @@ import at.ac.tuwien.kr.alpha.common.atoms.Literal;
 import at.ac.tuwien.kr.alpha.common.rule.BasicRule;
 import at.ac.tuwien.kr.alpha.common.terms.ArithmeticTerm;
 import at.ac.tuwien.kr.alpha.common.terms.ArithmeticTerm.ArithmeticOperator;
+import at.ac.tuwien.kr.alpha.grounder.transformation.aggregates.AggregateOperatorNormalization;
 import at.ac.tuwien.kr.alpha.common.terms.ConstantTerm;
 import at.ac.tuwien.kr.alpha.common.terms.Term;
 import at.ac.tuwien.kr.alpha.test.util.RuleParser;
